requirements.txt
CherryPy==3.2.0
PyYAML==3.10
RPyC==3.1.0
distribute==0.6.15
netaddr==0.7.5
wsgiref==0.1.2
